requirements.txt
# Run2Art — GPS Art Running Route Generator
# Core dependencies (required)
numpy>=1.24
scipy>=1.10
networkx>=3.1
osmnx>=1.6
pillow>=9.5
shapely>=2.0

# Optimizer (used by engine.py CMA-ES mode)
cma>=3.3

# Web server (Flask alternative to Node.js)
Flask>=3.0
requests>=2.28

# ── Optional dependencies (gracefully degraded if missing) ──
# scikit-image      # v8.4 skeleton scoring (use_skeleton_score flag)
# POT               # v8.4 Fused Gromov-Wasserstein (use_fgw flag)
# gudhi             # v8.4 persistent homology (use_ph_topology flag)
# torch             # v8.4 perceptual loss (use_perceptual_loss flag)
# torchvision       # v8.4 perceptual loss (use_perceptual_loss flag)
